refactor(home): ajusta e componentiza header
