+++ front-end/src/components/home/HeaderSlider.js
