+++ front-end/src/styles/home.css
@@ -1,8 +1,20 @@
-.header-home{
+.header-home {
     background-image: url('/images/imgBaseLeft.svg');
     background-repeat: no-repeat;
     background-size: cover;
     background-position: center;
+    position: relative;
+}
+
+.header-home::after {
+    content: "";
+    position: absolute;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    height: 30%;
+    pointer-events: none;
+    background: linear-gradient(to top, rgba(255,255,255,0.8) 0%, rgba(255,255,255,0) 100%);
 }
 
 .nav-link{
@@ -84,4 +96,20 @@
 
 .lemon-milk{
     font-family: 'Lemon Milk', sans-serif;
+}
+
+.cta-card-button{
+    background-color: var(--primary);
+    color: white;
+    border: none;
+    padding: 18px 84px;
+    width: fit-content;
+    font-weight: 500;
+    transition: background-color 0.3s ease, filter 0.3s ease;
+}
+
+.cta-card-button:hover{
+    background-color: var(--primary)!important;
+    color: white !important;
+    filter: brightness(90%);
 }
